Add run.sh and upd gitignore for gbad
This works! With the full 963k triple datastore
which I exported from GraphDB (entire Description,
Authority, and some smalle named graphs),
which is 174 MB in size in TriG, takes ca. 30secs
to start on my 2-core old laptop with 16 GB RAM,
running under Docker and under a reverse proxy!

Which means that probably the smallest VPS
will do for starters!